+++ voicelint/checkers/dyn_checker.py
@@ -24,6 +24,12 @@ class DynStrChecker(BaseChecker):
             "dyn-separated-strings",
             ""
         ),
+        "E11":(
+            '%s does not support %s syntax! You should probably refactor this into something like ' + 
+            '`Function(lambda %s: %s(%s).execute())`',
+            "no-dyn-strings",
+            ""
+        )
     }
 
 
@@ -37,3 +43,18 @@ class DynStrChecker(BaseChecker):
             text = '{}("{}")'.format(name,re.sub(r",{2,}",r",",js))
             self.add_message("dyn-separated-strings", node=node,args=(name,text))
 
+        if name in ["BringApp","StartApp"]:
+            if any(x for x in node.args if is_string(x) and re.fullmatch(r"%\((\w+)\)s",x.value)):
+                temporary,mistakes = [],[]
+                for a in node.args:
+                    if is_string(a):
+                        m = re.fullmatch(r"%\((\w+)\)s",a.value)
+                        if m:
+                            mistakes.append(m.group(1))
+                            temporary.append(m.group(1) + ".format()")
+                        else:
+                            temporary.append(a.as_string())
+                    else:
+                        temporary.append(a.as_string())
+                args = (name,"%(" + mistakes[0] + ")s",",".join(mistakes),name,",".join(temporary))
+                self.add_message("no-dyn-strings", node=node,args=args)