+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add check to catch errors with using multiple strings specs in Key,Text,Mouse action,e.g. `Key("c-a","c-c")` vs `Key("c-a,c-c")`
 
+- add checks so that `BringApp`,`StartApp` do not use dyn-strings and suggest refactoring
 
 ## Changed